Consilidated test_lfunction.py and test_lfunctions.py
